requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.5.0
attrs==21.2.0
Babel==2.8.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
daemonize==2.5.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==4.0.4
empy==3.3.4
filelock==3.6.0
flake8==4.0.1
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
gcovr==5.0
geocoder==1.38.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
intelhex==2.3.0
jeepney==0.7.1
Jinja2==3.0.3
keyboard==0.13.5
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
MarkupSafe==2.0.1
matplotlib==3.5.1
MAVProxy==1.8.49
mccabe==0.6.1
meson==0.61.2
more-itertools==8.10.0
mpmath==0.0.0
netifaces==0.11.0
notify2==0.3
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paho-mqtt==1.6.1
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.2
ply==3.11
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
pycodestyle==2.8.0
pycups==2.0.1
pyflakes==2.4.0
pygame==2.1.2
Pygments==2.11.2
PyGObject==3.42.0
PyJWT==2.3.0
pymacaroons==0.13.0
pymavlink==2.4.29
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyserial==3.5
PySticks==0.1
python-apt==2.3.0+ubuntu2
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
pythran==0.10.0
pytz==2022.1
pyudev==0.22.0
pyxdg==0.27
PyYAML==5.4.1
ratelim==0.1.6
reportlab==3.6.8
requests==2.25.1
scipy==1.8.0
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
sympy==1.9
systemd-python==234
tk==0.1.0
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
virtualenv==20.14.1
wadllib==1.3.6
webencodings==0.5.1
wxPython==4.0.7
xdg==5
xkit==0.0.0
zipp==1.0.0
